Updated All Classes
Updated View to properly handle user input and created MusicStore. The view functions properly for all implemented functions, but the ability to edit the user library still needs to be implemented. Code will be cleaned up and commented today, just wanted to get this in since it all at least works right now.